refactor #117: 나의 채팅탭 unread_count 컬럼 수정

+++ src/main/java/fastcampus/team7/Livable_officener/dto/ChatRoomListResponseDTO.java
@@ -17,7 +17,7 @@ public class ChatRoomListResponseDTO {
         private Long chatRoomId;
         private String storeName;
         private String recentMessage;
-        private Long numUnreadMessages;
+        private Integer numUnreadMessages;
         private FoodTag foodTag;
 
         public static ChatRoomListDTO of(ChatRoomListDTO chat) {
